ierg/requirements.txt
-e git://github.com/earle/django-bootstrap.git@ebe553a0b7f023124f334f2f5d45bd252220b5fe#egg=bootstrap
south
-e git://github.com/bradwhittington/django-cerial.git#egg=cerial
django-grappelli
redis>2.4.5
django-redis-cache
xlwt
xlrd
tablib
ordereddict
django-guardian
openpyxl
